Validate vdf steps using cached ones if are available
