+++ crates/actors/src/block_validation.rs
@@ -1,11 +1,12 @@
 use crate::{
     block_index_service::BlockIndexReadGuard, epoch_service::PartitionAssignmentsReadGuard,
-    mining::hash_to_number, vdf_service::VdfStepsReadGuard,
+    mining::hash_to_number,
 };
 use base58::ToBase58;
 use irys_database::Ledger;
 use irys_packing::{capacity_single::compute_entropy_chunk, xor_vec_u8_arrays_in_place};
 use irys_storage::ii;
+use irys_vdf::vdf_state::VdfStepsReadGuard;
 use irys_types::{
     calculate_difficulty, next_cumulative_diff, storage_config::StorageConfig, validate_path,
     Address, DifficultyAdjustmentConfig, IrysBlockHeader, PoaData, VDFStepsConfig, H256,
