+++ crates/types/configs/testnet.toml
@@ -3,8 +3,8 @@ difficulty_adjustment_interval = 2000
 max_difficulty_adjustment_factor = 4.0
 min_difficulty_adjustment_factor = 0.1
 chunk_size = 262_144
-num_chunks_in_partition = 51_872_000
-num_chunks_in_recall_range = 800
+num_chunks_in_partition = 800
+num_chunks_in_recall_range = 100
 vdf_reset_frequency = 1200
 vdf_parallel_verification_thread_limit = 4
 num_checkpoints_in_vdf_step = 25
