+++ crates/vdf/Cargo.toml
@@ -24,6 +24,7 @@ tracing.workspace = true
 tracing-subscriber.workspace = true
 tempfile.workspace = true
 openssl.workspace = true
+nodit.workspace = true
 
 [lints]
 workspace = true
